refactor: ai-core 라이브러리 삭제, AI관련 config 및 chatResult는 common으로 이동, root gradle에 AI 의존성 추가
