+++ ai-core/src/main/java/com/dev_high/product/ai/infrastructure/SpringAiChatModel.java
@@ -1,8 +1,6 @@
 package com.dev_high.product.ai.infrastructure;
 
 import com.dev_high.product.ai.domain.ChatMessage;
-import com.dev_high.product.ai.domain.ChatModel;
-import com.dev_high.product.ai.domain.ChatResult;
 import com.dev_high.product.ai.dto.DraftTonePreset;
 import com.dev_high.product.ai.dto.ProductDetailDto;
 import lombok.extern.slf4j.Slf4j;
@@ -23,56 +21,19 @@ import java.util.Map;
 
 @Component
 @Slf4j
-public class SpringAiChatModel implements ChatModel {
+public class SpringAiChatModel {
 
-//	private final ObjectProvider<ChatClient> chatClientProvider;
 	private final ChatClient chatClient;
 
-	private final PromptTemplate recommendTemplate;
     private final PromptTemplate imageToDetailWithCategoryTemplate;
 
 
-	@Value("${spring.ai.openai.api-key:}")
-	private String apiKey;
-
 	public SpringAiChatModel(
 			ChatClient chatClient,
-			@Qualifier("recommendTemplate") PromptTemplate recommendTemplate,
-            @Qualifier("imageToDetailWithCategoryTemplate") PromptTemplate imageToDetailWithCategoryTemplate
+			@Qualifier("imageToDetailWithCategoryTemplate") PromptTemplate imageToDetailWithCategoryTemplate
             ) {
 		this.chatClient = chatClient;
-		this.recommendTemplate = recommendTemplate;
         this.imageToDetailWithCategoryTemplate=imageToDetailWithCategoryTemplate;
-	}
-
-
-	@Override
-	public ChatResult chat(ChatMessage message) {
-
-//		ChatClient chatClient = chatClientProvider.getIfAvailable();
-		if (chatClient == null || apiKey == null || apiKey.isBlank() || "changeme".equals(apiKey)) {
-			log.warn("chat skipped: OpenAI API key not configured");
-			return new ChatResult("OpenAI API key not configured", Map.of("source", "fallback"));
-		}
-
-		log.info("chat request: question='{}', context='{}'",
-				message.value(),
-				message.context());
-		Prompt prompt = recommendTemplate.create(Map.of(
-				"question", message.value(),
-				"context", message.context()
-		));
-		log.info("chat prompt: {}", prompt);
-		ChatResponse response = chatClient.prompt(prompt).call().chatResponse();
-		Generation generation = response.getResult();
-		log.info("chat response raw: {}", response);
-
-		Map<String, Object> metadata = new HashMap<>();
-		response.getMetadata().entrySet()
-				.forEach(entry -> metadata.put(entry.getKey(), entry.getValue()));
-		String content = generation.getOutput().getText();
-		log.info("chat response: content='{}', metadata={}", content, metadata);
-		return new ChatResult(content, metadata);
 	}
 
     public ProductDetailDto chatProductDetail(
@@ -80,9 +41,6 @@ public class SpringAiChatModel implements ChatModel {
             String categoryOptions,   // ✅ 호출자가 만들어서 넘김
             Integer retryCount
     ) {
-        if (!isEnabled()) {
-            throw new IllegalStateException("OpenAI API key not configured");
-        }
 
 
         String systemText = imageToDetailWithCategoryTemplate.render(Map.of(
@@ -118,9 +76,6 @@ public class SpringAiChatModel implements ChatModel {
                 .entity(ProductDetailDto.class);
     }
 
-    private boolean isEnabled() {
-        return !(chatClient == null || apiKey == null || apiKey.isBlank() || "changeme".equals(apiKey));
-    }
 
     private void validateImage(MultipartFile file) {
         if (file == null || file.isEmpty()) throw new IllegalArgumentException("file is empty");
