+++ build.gradle
@@ -47,6 +47,7 @@ subprojects {
     dependencyManagement {
         imports {
             mavenBom "org.springframework.cloud:spring-cloud-dependencies:$springCloudVersion"
+            mavenBom "org.springframework.ai:spring-ai-bom:1.1.2"
         }
     }
 
@@ -89,6 +90,8 @@ subprojects {
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
         testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
+        //AI
+        implementation 'org.springframework.ai:spring-ai-starter-model-openai'
 
     }
