+++ product/src/main/java/com/dev_high/product/application/ProductRecommendService.java
@@ -1,30 +1,38 @@
 package com.dev_high.product.application;
 
 import com.dev_high.product.ai.domain.ChatMessage;
-import com.dev_high.product.ai.domain.ChatModel;
-import com.dev_high.product.ai.domain.ChatResult;
+import com.dev_high.common.dto.ChatResult;
 import com.dev_high.product.application.dto.ProductAnswer;
 import com.dev_high.product.application.dto.ProductSearchInfo;
 import com.dev_high.product.domain.Category;
 import com.dev_high.product.domain.Product;
 import com.dev_high.product.domain.ProductRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.ai.chat.client.ChatClient;
+import org.springframework.ai.chat.model.ChatResponse;
+import org.springframework.ai.chat.model.Generation;
+import org.springframework.ai.chat.prompt.Prompt;
+import org.springframework.ai.chat.prompt.PromptTemplate;
 import org.springframework.ai.document.Document;
 import org.springframework.ai.vectorstore.SearchRequest;
 import org.springframework.ai.vectorstore.VectorStore;
 import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 @RequiredArgsConstructor
 public class ProductRecommendService {
 
     private final ProductRepository productRepository;
     private final VectorStore vectorStore;
-    private final ChatModel chatModel;
+    private final ChatClient chatClient;
+    private final PromptTemplate recommendTemplate;
 
 
 
@@ -84,11 +92,25 @@ public class ProductRecommendService {
 
         List<ProductSearchInfo> productSearchInfo= search(query, topK);
         String context = toContext(productSearchInfo);
-        ChatResult result = chatModel.chat(new ChatMessage(query, context));
 
-        ProductAnswer answer=new ProductAnswer(result.content(), productSearchInfo);
+        Prompt prompt = recommendTemplate.create(Map.of(
+                "question", query,
+                "context", context
+        ));
 
-        return answer;
+        ChatResponse response = chatClient.prompt(prompt).call().chatResponse();
+        log.info("chat response raw: {}", response);
+
+        Map<String, Object> metadata = new HashMap<>();
+        response.getMetadata().entrySet()
+                .forEach(entry -> metadata.put(entry.getKey(), entry.getValue()));
+
+        String content = response.getResult().getOutput().getText();
+        log.info("chat response: content='{}', metadata={}", content, metadata);
+
+        ChatResult<String> result = new ChatResult<>(content, metadata);
+
+        return new ProductAnswer(result.content(), productSearchInfo);
     }
